+++ arctic/_util.py
@@ -12,20 +12,8 @@ logger = logging.getLogger(__name__)
 
 NP_OBJECT_DTYPE = np.dtype('O')
 
-# Avoid import-time extra logic
-_use_new_count_api: bool | None = None
-
-
 def get_fwptr_config(version: dict[str, Any]) -> FwPointersCfg:
     return FwPointersCfg[version.get(FW_POINTERS_CONFIG_KEY, FwPointersCfg.DISABLED.name)]
-
-
-def _detect_new_count_api() -> bool:
-    try:
-        mongo_v = [int(v) for v in pymongo.version.split('.')]
-        return mongo_v[0] >= 3 and mongo_v[1] >= 7
-    except:
-        return False
 
 
 def indent(s: str, num_spaces: int) -> str:
@@ -84,16 +72,8 @@ def mongo_count(collection: Any, filter: dict[str, Any] | None = None, **kwargs:
     use with care as filters on un-indexed fields will generate COLLSCAN.
     """
     filter = {} if filter is None else filter
-    global _use_new_count_api
-    _use_new_count_api = _detect_new_count_api() if _use_new_count_api is None else _use_new_count_api
-
-    if _use_new_count_api:
-        if filter == {}:
-            # fast. uses collection metadata
-            return cast(int, collection.estimated_document_count(**kwargs))
-        else:
-            # transactions supported, but slow for non-indexed filters
-            return cast(int, collection.count_documents(filter=filter, **kwargs))
-    else:
-        # pymongo <= 3.6 # faster than count_documents but non-transactional and deprecated
-        return cast(int, collection.count(filter=filter, **kwargs))
+    if filter == {}:
+        # Fast: uses collection metadata.
+        return cast(int, collection.estimated_document_count(**kwargs))
+    # Transactions supported, but slow for non-indexed filters.
+    return cast(int, collection.count_documents(filter=filter, **kwargs))